test/Simulation factored simulated data generation from test_Analyse.cpp
