+++ test/CMakeLists.txt
@@ -30,6 +30,7 @@
 # === Test Programs
 # ===
 
+set(testLibName "OriManiaTest")
 
 set(mainProgs
 
@@ -41,6 +42,27 @@ set(mainProgs
 	test_Convention # diverse conventions for representing orientations
 	test_Orientation # math operations involving orientation data
 
+	)
+
+add_library(${testLibName}
+	
+	Simulation.cpp
+
+	)
+
+target_include_directories(
+	${testLibName}
+	PUBLIC
+		${CMAKE_CURRENT_SOURCE_DIR}/../include # public interface
+	PRIVATE
+		${CMAKE_CURRENT_SOURCE_DIR}  # local test code includes
+	)
+
+target_link_libraries(
+	${testLibName}
+	PRIVATE
+		Engabra::Engabra
+		Rigibra::Rigibra
 	)
 
 foreach(mainProg ${mainProgs})
@@ -68,6 +90,7 @@ foreach(mainProg ${mainProgs})
 		PRIVATE
 			Engabra::Engabra
 			Rigibra::Rigibra
+			${testLibName}
 			${aProjLib}
 		)
